+++ @xen-orchestra/lite/src/stores/xen-api/host.store.ts
@@ -4,6 +4,7 @@ import type { XenApiHost } from "@/libs/xen-api/xen-api.types";
 import { useXenApiStore } from "@/stores/xen-api.store";
 import { createUseCollection } from "@/stores/xen-api/create-use-collection";
 import { useHostMetricsStore } from "@/stores/xen-api/host-metrics.store";
+import type { XenApiPatch } from "@/types/xen-api";
 import { defineStore } from "pinia";
 import { computed } from "vue";
 
@@ -42,10 +43,36 @@ export const useHostStore = defineStore("xen-api-host", () => {
     });
   }) as GetStats<XenApiHost>;
 
+  const fetchMissingPatches = async (
+    hostRef: XenApiHost["$ref"]
+  ): Promise<XenApiPatch[]> => {
+    const xenApiStore = useXenApiStore();
+
+    const rawPatchesAsString = await xenApiStore
+      .getXapi()
+      .call<string>("host.call_plugin", [
+        hostRef,
+        "updater.py",
+        "check_update",
+        {},
+      ]);
+
+    const rawPatches = JSON.parse(rawPatchesAsString) as Omit<
+      XenApiPatch,
+      "$id"
+    >[];
+
+    return rawPatches.map((rawPatch) => ({
+      ...rawPatch,
+      $id: `${rawPatch.name}-${rawPatch.version}`,
+    }));
+  };
+
   return {
     ...context,
     runningHosts,
     getStats,
+    fetchMissingPatches,
   };
 });