+++ @xen-orchestra/lite/src/types/xen-api.ts
@@ -21,3 +21,19 @@ export interface XenApiAlarm<RelationType extends RawObjectType>
   triggerLevel: number;
   type: XenApiAlarmType;
 }
+
+export type XenApiPatch = {
+  $id: string;
+  name: string;
+  description: string;
+  license: string;
+  release: string;
+  size: number;
+  url: string;
+  version: string;
+  changelog: {
+    date: number;
+    description: string;
+    author: string;
+  };
+};